Modification in the code

+++ src/main/java/br/unicamp/ctm/representation/idea/MatrixIdeaMetadataValues.java
@@ -6,37 +6,37 @@ import scala.Char;
 
 public class MatrixIdeaMetadataValues {
 
-  public static Map<Class, Double> getMetadataMap() {
+  public static Map<Class, Integer> getMetadataMap() {
 
-    HashMap<Class, Double> metadataMap = new HashMap<>();
+    HashMap<Class, Integer> metadataMap = new HashMap<>();
 
-    metadataMap.put(int.class, 1d);
-    metadataMap.put(Integer.class, 1d);
-    metadataMap.put(float.class, 2d);
-    metadataMap.put(Float.class, 2d);
-    metadataMap.put(double.class, 3d);
-    metadataMap.put(Double.class, 2d);
-    metadataMap.put(char.class, 4d);
-    metadataMap.put(Char.class, 4d);
-    metadataMap.put(short.class, 5d);
-    metadataMap.put(Short.class, 5d);
-    metadataMap.put(boolean.class, 6d);
-    metadataMap.put(Boolean.class, 6d);
-    metadataMap.put(String.class, 7d);
+    metadataMap.put(int.class, 1);
+    metadataMap.put(Integer.class, 1);
+    metadataMap.put(float.class, 2);
+    metadataMap.put(Float.class, 2);
+    metadataMap.put(double.class, 3);
+    metadataMap.put(Double.class, 2);
+    metadataMap.put(char.class, 4);
+    metadataMap.put(Char.class, 4);
+    metadataMap.put(short.class, 5);
+    metadataMap.put(Short.class, 5);
+    metadataMap.put(boolean.class, 6);
+    metadataMap.put(Boolean.class, 6);
+    metadataMap.put(String.class, 7);
 
-    metadataMap.put(int[].class, 8d);
-    metadataMap.put(Integer[].class, 8d);
-    metadataMap.put(double[].class, 9d);
-    metadataMap.put(Double[].class, 9d);
-    metadataMap.put(float[].class, 10d);
-    metadataMap.put(Float[].class, 10d);
-    metadataMap.put(short[].class, 11d);
-    metadataMap.put(Short[].class, 11d);
-    metadataMap.put(long[].class, 12d);
-    metadataMap.put(Long[].class, 12d);
-    metadataMap.put(boolean[].class, 13d);
-    metadataMap.put(Boolean[].class, 13d);
-    metadataMap.put(String[].class, 14d);
+    metadataMap.put(int[].class, 8);
+    metadataMap.put(Integer[].class, 8);
+    metadataMap.put(double[].class, 9);
+    metadataMap.put(Double[].class, 9);
+    metadataMap.put(float[].class, 10);
+    metadataMap.put(Float[].class, 10);
+    metadataMap.put(short[].class, 11);
+    metadataMap.put(Short[].class, 11);
+    metadataMap.put(long[].class, 12);
+    metadataMap.put(Long[].class, 12);
+    metadataMap.put(boolean[].class, 13);
+    metadataMap.put(Boolean[].class, 13);
+    metadataMap.put(String[].class, 14);
 
     return metadataMap;
   }
